Various Improvements To All Files
Modified and added all the files and added more customization. Added one new file, dat_analysis.py, which uses t-sne to visualize how the network is classifying different states. Fixed part of memory problems, allowing for more states to be stored in memory for training. Tried to make exiting of a run more graceful. Tweaked the Configuration settings a lot.